Add Pacman animation. Refactor bitmaps.
Added the Pacman animation of the ghost being chased by pacman moving to the left side.

Refactored the bitmaps down to fewer variables, thus creating multidimentional arrays. This makes for slightly easier usage of the bitmaps.

Added a frame counter rather than straight modulo to get the frame. There are subtle bugs when taking the modulo of negative numbers in C/C++.